requirements.txt
aiohttp==3.13.3
discord.py==2.7.1
python-dotenv==1.2.2
